@@ -198,105 +178,105 @@
         "lines": 98
       }
     },
-    "moduleDirectories": [
-      "node_modules",
-      "app"
-    ],
+    "moduleDirectories": ["node_modules", "app"],
     "moduleNameMapper": {
-      ".*\\.(css|less|styl|scss|sass)$": "<rootDir>/internals/mocks/cssModule.js",
-      ".*\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/internals/mocks/image.js"
+      ".*\\.(css|less|styl|scss|sass)$":
+        "<rootDir>/internals/mocks/cssModule.js",
+      ".*\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$":
+        "<rootDir>/internals/mocks/image.js"
     },
-    "setupTestFrameworkScriptFile": "<rootDir>/internals/testing/test-bundler.js",
+    "setupTestFrameworkScriptFile":
+      "<rootDir>/internals/testing/test-bundler.js",
     "testRegex": "tests/.*\\.test\\.js$"
   },
   "dependencies": {
-    "axios": "^0.16.2",
-    "babel-polyfill": "6.20.0",
-    "chalk": "^1.1.3",
-    "compression": "1.6.2",
-    "cross-env": "3.1.3",
-    "express": "4.14.0",
-    "fontfaceobserver": "2.0.7",
+    "axios": "0.16.2",
+    "babel-polyfill": "6.26.0",
+    "chalk": "^2.1.0",
+    "compression": "1.7.0",
+    "cross-env": "5.0.5",
+    "express": "4.15.4",
+    "fontfaceobserver": "2.0.13",
     "immutable": "3.8.1",
     "intl": "1.2.5",
     "invariant": "2.2.2",
-    "ip": "1.1.4",
-    "lodash": "4.17.2",
+    "ip": "1.1.5",
+    "lodash": "4.17.4",
     "minimist": "1.2.0",
-    "node-upwork": "^0.0.7",
-    "react": "15.4.1",
-    "react-burger-menu": "2.1.5",
-    "react-dom": "15.4.1",
-    "react-helmet": "3.2.2",
-    "react-intl": "2.1.5",
-    "react-redux": "4.4.6",
+    "node-upwork": "0.0.7",
+    "react": "15.6.1",
+    "react-burger-menu": "2.1.6",
+    "react-dom": "15.6.1",
+    "react-helmet": "5.2.0",
+    "react-intl": "2.4.0",
+    "react-redux": "5.0.6",
     "react-router": "3.0.0",
     "react-router-redux": "4.0.6",
     "react-router-scroll": "0.4.1",
     "react-scroll": "1.5.4",
     "react-text-loop": "1.1.0",
-    "redux": "3.6.0",
-    "redux-immutable": "3.0.8",
-    "redux-saga": "0.14.0",
-    "reselect": "2.5.4",
-    "sanitize.css": "4.1.0",
-    "styled-components": "^2.1.2",
+    "redux": "3.7.2",
+    "redux-immutable": "4.0.0",
+    "redux-saga": "0.15.6",
+    "reselect": "3.0.1",
+    "sanitize.css": "5.0.0",
+    "styled-components": "2.1.2",
     "superagent": "3.6.0",
     "upwork-api": "1.2.3",
     "warning": "3.0.0",
-    "whatwg-fetch": "2.0.1"
+    "whatwg-fetch": "2.0.3"
   },
   "devDependencies": {
-    "babel-cli": "6.18.0",
-    "babel-core": "6.21.0",
-    "babel-eslint": "7.1.1",
-    "babel-loader": "6.2.10",
+    "babel-cli": "6.26.0",
+    "babel-core": "6.26.0",
+    "babel-eslint": "7.2.3",
+    "babel-loader": "7.1.2",
     "babel-plugin-dynamic-import-node": "1.0.0",
-    "babel-plugin-react-intl": "2.2.0",
+    "babel-plugin-react-intl": "2.3.1",
     "babel-plugin-react-transform": "2.0.2",
-    "babel-plugin-transform-es2015-modules-commonjs": "6.18.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "6.26.0",
     "babel-plugin-transform-react-constant-elements": "6.9.1",
     "babel-plugin-transform-react-inline-elements": "6.8.0",
     "babel-plugin-transform-react-remove-prop-types": "0.2.11",
-    "babel-preset-latest": "6.16.0",
-    "babel-preset-react": "6.16.0",
+    "babel-preset-latest": "6.24.1",
+    "babel-preset-react": "6.24.1",
     "babel-preset-react-hmre": "1.1.1",
-    "babel-preset-stage-0": "6.16.0",
+    "babel-preset-stage-0": "6.24.1",
     "cheerio": "0.22.0",
-    "circular-dependency-plugin": "2.0.0",
-    "coveralls": "2.11.15",
-    "css-loader": "0.26.1",
-    "enzyme": "2.6.0",
-    "eslint": "3.11.1",
-    "eslint-config-airbnb": "13.0.0",
-    "eslint-config-airbnb-base": "10.0.1",
-    "eslint-import-resolver-webpack": "0.8.0",
-    "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.7.1",
-    "eslint-plugin-redux-saga": "0.1.5",
+    "circular-dependency-plugin": "4.2.0",
+    "coveralls": "2.13.1",
+    "css-loader": "0.28.7",
+    "enzyme": "2.9.1",
+    "eslint": "4.6.1",
+    "eslint-config-airbnb": "15.1.0",
+    "eslint-config-airbnb-base": "12.0.0",
+    "eslint-import-resolver-webpack": "0.8.3",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-jsx-a11y": "6.0.2",
+    "eslint-plugin-react": "7.3.0",
+    "eslint-plugin-redux-saga": "0.4.0",
     "eventsource-polyfill": "0.9.6",
-    "exports-loader": "0.6.3",
-    "file-loader": "0.9.0",
-    "html-loader": "0.4.4",
+    "exports-loader": "0.6.4",
+    "file-loader": "0.11.2",
+    "html-loader": "0.5.1",
     "html-webpack-plugin": "2.24.1",
     "image-webpack-loader": "2.0.0",
     "imports-loader": "0.6.5",
-    "jest-cli": "18.0.0",
-    "lint-staged": "3.2.1",
+    "jest-cli": "21.0.2",
+    "lint-staged": "4.1.3",
     "ngrok": "2.2.4",
     "node-plop": "0.5.4",
     "null-loader": "0.1.1",
     "offline-plugin": "4.5.2",
     "plop": "1.7.3",
     "pre-commit": "1.1.3",
-    "prop-types": "^15.5.10",
+    "prop-types": "15.5.10",
     "react-addons-test-utils": "15.4.1",
-    "rimraf": "2.5.4",
+    "rimraf": "2.6.2",
     "shelljs": "^0.7.5",
-    "sinon": "2.0.0-pre",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
+    "sinon": "3.2.1",
+    "style-loader": "0.18.2",
+    "url-loader": "0.5.9",
     "webpack": "2.2.0-rc.3",
     "webpack-dev-middleware": "1.9.0",
     "webpack-hot-middleware": "2.15.0"
