@@ -8,7 +8,12 @@ const CallToActionButton = styled(Link)`
   color: inherit;
   background: ${(props) => (props.background ? props.background : 'none')};
   cursor: pointer;
-  padding: ${(props) => (props.medium ? '8px 50px' : '25px 80px')};
+
+  padding-top: ${(props) => (props.paddingTop ? props.paddingTop : '25px')};
+  padding-bottom: ${(props) =>
+    props.paddingBottom ? props.paddingBottom : '25px'};
+  padding-left: ${(props) => (props.paddingLeft ? props.paddingLeft : '80px')};
+  padding-right: ${(props) => (props.paddingRight ? props.paddingRight : '80px')};
   display: inline-block;
   text-transform: uppercase;
   letter-spacing: 1px;
@@ -18,7 +23,7 @@ const CallToActionButton = styled(Link)`
   transition: all 0.3s;
   text-decoration: none;
   border: 3px solid #fff;
-  color: ${(props) => (props.background ? '#5cbb2f' : '#fff')};
+  color: ${(props) => (props.color ? props.color : '#fff')};
 
   &:after {
     content: '';
@@ -27,10 +32,10 @@ const CallToActionButton = styled(Link)`
     transition: all 0.3s;
   }
 
-
   &:hover {
     color: #5cbb2f;
     background: ${(props) => (props.background ? '#f4f8f1' : '#fff')};
+    border-color: ${(props) => (props.background ? '#5cbb2f' : '#fff')};
   }
 
   &:active {
