@@ -1,28 +1,45 @@
 import React from 'react';
 import Proptypes from 'prop-types';
+
+import A from 'components/A';
 import ListItem from './styles/listItem';
 import Date from './styles/date';
 import TimeLine from './styles/timeline';
 import Info from './styles/info';
-
+import ListItemTitle from './ListItemTitle';
 
 const List = (props) => {
   const [startYear, , endYear] = props.year.split(' ');
   return (
     <ListItem>
       <div>
-        <Date right>{ startYear }</Date>
-        <Date iscolored bold large>{ endYear }</Date>
+        <Date right>{startYear}</Date>
+        <Date iscolored bold large>
+          {endYear}
+        </Date>
       </div>
       <TimeLine>
-        <span></span>
+        <span />
       </TimeLine>
       <Info>
-        <h4>{ props.school }</h4>
+        <h4>{props.school}</h4>
         <div>
-          <span>{ props.grade }</span>
-          <span>{ props.address }</span>
-          <span>{ props.course }</span>
+          <span>
+            <ListItemTitle>Grade</ListItemTitle>
+            {props.grade}
+          </span>
+          <span>
+            <ListItemTitle>Address</ListItemTitle>
+            {props.address}
+          </span>
+          <span>
+            <ListItemTitle>Course</ListItemTitle>
+            {props.course}
+          </span>
+          <span>
+            <ListItemTitle>Link</ListItemTitle>
+            <A href={props.website}>{props.website}</A>
+          </span>
         </div>
       </Info>
     </ListItem>
@@ -35,6 +52,7 @@ List.propTypes = {
   grade: Proptypes.string.isRequired,
   course: Proptypes.string.isRequired,
   address: Proptypes.string.isRequired,
+  website: Proptypes.string.isRequired,
 };
 
 export default List;
