@@ -4,21 +4,43 @@ import Helmet from 'react-helmet';
 import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
+import styled from 'styled-components';
 
 import { makeSelectGeneralData } from 'containers/App/selectors';
 import H2 from 'components/H2';
-import withProgressBar from 'components/ProgressBar';
+
+import GitHub2 from 'components/Icons/GitHub2';
+import Upwork from 'components/Icons/Upwork';
+import LinkedIn from 'components/Icons/LinkedIn';
+import WordPress from 'components/Icons/WordPress';
+import A from 'components/A';
 
 import CenteredSection from './CenteredSection';
 import Section from './Section';
 import messages from './messages';
 import { loadGeneralData } from '../App/actions';
 
-function Component() {
-  return <div />;
-}
+const ReferenceWrapper = styled.ul`
+  display: flex;
+  flex-wrap: wrap;
 
-const HocComponent = withProgressBar(Component);
+  margin: 1rem 0 0 0;
+  list-style: none;
+  text-align: center;
+`;
+
+const ReferenceItem = styled.li`
+  width: 25%;
+  height: 4em;
+  text-align: center;
+  transition: 0.3s;
+  transition-property: margin;
+  justify-content: center;
+
+  &:hover {
+    margin-top: : 1px;
+  }
+`;
 
 export class HomePage extends React.PureComponent {
   componentWillMount() {
@@ -52,53 +74,77 @@ export class HomePage extends React.PureComponent {
             <H2>Languages</H2>
             <ul>
               {general.languages &&
-                general.languages.map((item, index) =>
-                  <li key={index}>
+                general.languages.map((item) =>
+                  (<li key={item.id}>
                     {item.id} -
                     {item.name} -
                     {item.level} -
                     {item.comment} -
-                  </li>
+                  </li>)
                 )}
             </ul>
           </Section>
           <Section>
-            <H2>
-              <FormattedMessage {...messages.show_testimonial} />
-            </H2>
+            <H2>Awards</H2>
             <ul>
-              {general.references &&
-                Object.keys(general.references).map((item, index) =>
-                  <li key={index}>
-                    {general.references[item]}
-                  </li>
+              {general.awards &&
+                general.awards.map((item) =>
+                  (<li key={item.id}>
+                    {item.id} -
+                    {item.year}
+                    {item.title} -
+                  </li>)
                 )}
             </ul>
           </Section>
           <Section background="#f9f9f9">
-            <H2>Awards</H2>
+            <H2>Hobbies</H2>
             <ul>
-              {general.awards &&
-                general.awards.map((item, index) =>
-                  <li key={index}>
-                    {item.id} -
-                    {item.year}
-                    {item.title} -
-                  </li>
+              {general.interests &&
+                general.interests.map((item) =>
+                  (<li key={item}>
+                    {item}
+                  </li>)
                 )}
             </ul>
           </Section>
           <Section>
-            <H2>Hobbies</H2>
-            <ul>
-              {general.interests &&
-                general.interests.map((item, index) =>
-                  <li key={index}>
-                    {item}
-                  </li>
-                )}
-            </ul>
-            <HocComponent />
+            <ReferenceWrapper>
+              <ReferenceItem>
+                <A
+                  href={general.references && general.references.upwork}
+                  target="_blank"
+                >
+                  <Upwork />
+                </A>
+              </ReferenceItem>
+              <ReferenceItem>
+                <A
+                  href={general.references && general.references.linkedin}
+                  target="_blank"
+                >
+                  <LinkedIn />
+                </A>
+              </ReferenceItem>
+              <ReferenceItem>
+                <A
+                  href={
+                    general.references && general.references.wordpressPlugin
+                  }
+                  target="_blank"
+                >
+                  <WordPress />
+                </A>
+              </ReferenceItem>
+              <ReferenceItem>
+                <A
+                  href={general.references && general.references.github}
+                  target="_blank"
+                >
+                  <GitHub2 height="10em" width="10em" />
+                </A>
+              </ReferenceItem>
+            </ReferenceWrapper>
           </Section>
         </div>
       </article>
