@@ -1,53 +1,11 @@
+import { call, put, takeLatest } from 'redux-saga/effects';
+import { LOAD_GENERAL_DATA } from 'containers/App/constants';
 import {
-  take,
-  call,
-  put,
-  select,
-  cancel,
-  takeLatest,
-} from 'redux-saga/effects';
-import { LOCATION_CHANGE } from 'react-router-redux';
-import { LOAD_REPOS, LOAD_GENERAL_DATA } from 'containers/App/constants';
-import {
-  reposLoaded,
-  repoLoadingError,
   loadGeneralDataSuccss,
   loadGeneralDataFailure,
 } from 'containers/App/actions';
 
 import request from 'utils/request';
-import { makeSelectUsername } from 'containers/HomePage/selectors';
-
-/**
- * Github repos request/response handler
- */
-export function* getRepos() {
-  // Select username from store
-  const username = yield select(makeSelectUsername());
-  const requestURL = `https://api.github.com/users/${username}/repos?type=all&sort=updated`;
-
-  try {
-    // Call our request helper (see 'utils/request')
-    const repos = yield call(request, requestURL);
-    yield put(reposLoaded(repos, username));
-  } catch (err) {
-    yield put(repoLoadingError(err));
-  }
-}
-
-/**
- * Root saga manages watcher lifecycle
- */
-export function* githubData() {
-  // Watches for LOAD_REPOS actions and calls getRepos when one comes in.
-  // By using `takeLatest` only the result of the latest API call is applied.
-  // It returns task descriptor (just like fork) so we can continue execution
-  const watcher = yield takeLatest(LOAD_REPOS, getRepos);
-
-  // Suspend execution until location changes
-  yield take(LOCATION_CHANGE);
-  yield cancel(watcher);
-}
 
 export function* watchGeneralDataLoading() {
   yield takeLatest(LOAD_GENERAL_DATA, getGeneralData);
