@@ -1,3 +1,3 @@
-export const LOAD_EDUCATIONS = 'skmukhiya/App/LOAD_EDUCATIONS';
-export const LOAD_EDUCATIONS_SUCCESS = 'skmukhiya/App/LOAD_EDUCATIONS_SUCCESS';
-export const LOAD_EDUCATIONS_FAILURE = 'skmukhiya/App/LOAD_EDUCATIONS_ERROR';
+export const LOAD_REPOS = 'skmukhiya/App/LOAD_REPOS';
+export const LOAD_REPOS_SUCCESS = 'skmukhiya/App/LOAD_REPOS_SUCCESS';
+export const LOAD_REPOS_ERROR = 'skmukhiya/App/LOAD_REPOS_ERROR';
