@@ -1,33 +1,18 @@
-/**
- * RepoListItem
- *
- * Lists the name and the issue count of a repository
- */
-
-import React from 'react';
-import { connect } from 'react-redux';
-import { createStructuredSelector } from 'reselect';
+import React, { PureComponent } from 'react';
 import { FormattedNumber } from 'react-intl';
 
-import { makeSelectCurrentUser } from 'containers/App/selectors';
 import ListItem from 'components/ListItem';
 import IssueIcon from './IssueIcon';
 import IssueLink from './IssueLink';
 import RepoLink from './RepoLink';
 import Wrapper from './Wrapper';
 
-export class RepoListItem extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
+export class RepoListItem extends PureComponent {
+  // eslint-disable-line react/prefer-stateless-function
   render() {
     const item = this.props.item;
-    let nameprefix = '';
+    const nameprefix = '';
 
-    // If the repository is owned by a different person than we got the data for
-    // it's a fork and we should show the name of the owner
-    if (item.owner.login !== this.props.currentUser) {
-      nameprefix = `${item.owner.login}/`;
-    }
-
-    // Put together the content of the repository
     const content = (
       <Wrapper>
         <RepoLink href={item.html_url} target="_blank">
@@ -40,18 +25,12 @@ export class RepoListItem extends React.PureComponent { // eslint-disable-line r
       </Wrapper>
     );
 
-    // Render the content into a list item
-    return (
-      <ListItem key={`repo-list-item-${item.full_name}`} item={content} />
-    );
+    return <ListItem key={`repo-list-item-${item.full_name}`} item={content} />;
   }
 }
 
 RepoListItem.propTypes = {
   item: React.PropTypes.object,
-  currentUser: React.PropTypes.string,
 };
 
-export default connect(createStructuredSelector({
-  currentUser: makeSelectCurrentUser(),
-}))(RepoListItem);
+export default RepoListItem;
