@@ -1,20 +1,26 @@
-import React from 'react';
+import React, { Component } from 'react';
 import Helmet from 'react-helmet';
 import { FormattedMessage } from 'react-intl';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import { createStructuredSelector } from 'reselect';
 
 import Wrapper from 'components/PageContainer';
 import H1 from 'components/H1';
+
+import { makeSelectError } from 'containers/App/selectors';
+import { makeSelectSkills } from './selectors';
+
+import { loadSkills } from './actions';
 import messages from './messages';
 import List from './List';
 import ListItem from './ListItem';
 import ListItemTitle from './ListItemTitle';
+import Level from './Level';
 
-// eslint-disable-line react/prefer-stateless-function
-export default class SkillsPage extends React.Component {
-  // Since state and props are static,
-  // there's no need to re-render this component
-  shouldComponentUpdate() {
-    return false;
+class SkillsPage extends Component {
+  componentWillMount() {
+    this.props.loadSkills();
   }
 
   render() {
@@ -34,53 +40,44 @@ export default class SkillsPage extends React.Component {
             <FormattedMessage {...messages.header} />
           </H1>
           <List>
-            <ListItem>
-              <ListItemTitle>
-                <FormattedMessage {...messages.scaffoldingHeader} />
-              </ListItemTitle>
-              <p>
-                <FormattedMessage {...messages.scaffoldingMessage} />
-              </p>
-            </ListItem>
-
-            <ListItem>
-              <ListItemTitle>
-                <FormattedMessage {...messages.feedbackHeader} />
-              </ListItemTitle>
-              <p>
-                <FormattedMessage {...messages.feedbackMessage} />
-              </p>
-            </ListItem>
-
-            <ListItem>
-              <ListItemTitle>
-                <FormattedMessage {...messages.routingHeader} />
-              </ListItemTitle>
-              <p>
-                <FormattedMessage {...messages.routingMessage} />
-              </p>
-            </ListItem>
-
-            <ListItem>
-              <ListItemTitle>
-                <FormattedMessage {...messages.networkHeader} />
-              </ListItemTitle>
-              <p>
-                <FormattedMessage {...messages.networkMessage} />
-              </p>
-            </ListItem>
-
-            <ListItem>
-              <ListItemTitle>
-                <FormattedMessage {...messages.intlHeader} />
-              </ListItemTitle>
-              <p>
-                <FormattedMessage {...messages.intlMessage} />
-              </p>
-            </ListItem>
+            {this.props.skills &&
+              this.props.skills.map((item) => (
+                <ListItem key={item.id}>
+                  <ListItemTitle>
+                    {item.framework}
+                    {' - '}
+                    <Level>{item.level}</Level>
+                  </ListItemTitle>
+                  <p>{item.title}</p>
+                  <ul>
+                    {item.comments &&
+                      item.comments.map((comment, index) => (
+                        <li key={index}>{comment}</li>
+                      ))}
+                  </ul>
+                </ListItem>
+              ))}
           </List>
         </Wrapper>
       </div>
     );
   }
 }
+
+SkillsPage.propTypes = {
+  skills: PropTypes.oneOfType([PropTypes.array, PropTypes.bool]),
+  loadSkills: PropTypes.func.isRequired,
+};
+
+export function mapDispatchToProps(dispatch) {
+  return {
+    loadSkills: () => dispatch(loadSkills()),
+  };
+}
+
+const mapStateToProps = createStructuredSelector({
+  error: makeSelectError(),
+  skills: makeSelectSkills(),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(SkillsPage);
